requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
dnspython==2.7.0
idna==3.10
mysql-connector-python==9.3.0
pymongo==4.12.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
python-decouple==3.8
